RAG/requirements.txt
accelerate==0.32.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-search-documents==11.4.0b11
azure-storage-blob==12.23.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
bpemb==0.3.5
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
conllu==5.0.1
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
flair==0.13.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.2.0
fuzzywuzzy==0.18.0
gdown==5.2.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.23.5
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
Janome==0.5.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.3.4
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.137
lazy-object-proxy==1.10.0
Levenshtein==0.25.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
motor==3.6.0
mpld3==0.5.10
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
narwhals==1.11.1
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.52.2
openapi-core==0.19.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.5
orjson==3.10.10
packaging==24.1
pandas==2.2.2
parse==1.20.2
parso==0.8.4
pathable==0.4.3
pillow==10.4.0
plantuml==0.3.0
platformdirs==4.2.2
portalocker==2.10.1
pptree==3.1
prance==23.6.21.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==5.27.2
psutil==6.0.0
pure-eval==0.2.2
pyarrow==18.0.0
pybars4==0.9.13
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyMeta3==0.5.1
pymongo==4.9.2
PyMuPDF==1.24.12
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytorch_revgrad==0.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.4
rdflib==7.0.0
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.3
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.10.1
segtok==1.5.11
semantic-kernel==0.9.3b1
semver==3.0.2
sentencepiece==0.2.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.36
sqlitedict==2.1.0
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.0
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.0.6
Wikipedia-API==0.6.0
wrapt==1.16.0
yarl==1.16.0
